rename stl.h to vector.h
