+++ cpp/src/arrow/record_batch.cc
@@ -32,7 +32,7 @@
 #include "arrow/util/atomic_shared_ptr.h"
 #include "arrow/util/iterator.h"
 #include "arrow/util/logging.h"
-#include "arrow/util/stl.h"
+#include "arrow/util/vector.h"
 
 namespace arrow {
