+++ cpp/src/arrow/table.cc
@@ -32,7 +32,7 @@
 #include "arrow/type.h"
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/logging.h"
-#include "arrow/util/stl.h"
+#include "arrow/util/vector.h"
 
 namespace arrow {
