+++ cpp/src/arrow/type.cc
@@ -34,7 +34,7 @@
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/key_value_metadata.h"
 #include "arrow/util/logging.h"
-#include "arrow/util/stl.h"
+#include "arrow/util/vector.h"
 #include "arrow/visitor_inline.h"
 
 namespace arrow {
