+++ cpp/src/arrow/util/sort.h
@@ -71,4 +71,3 @@ size_t Permute(const std::vector<int64_t>& indices, std::vector<T>* values) {
 
 }  // namespace internal
 }  // namespace arrow
-
