+++ cpp/src/arrow/util/vector.h
